cnf ran: cnf/ran/internal reorganization
* Moves version testing into its own package.
* Creates a new rancluster package for cluster-level utilities, such as checking if it is SNO+1.
* Updates the test suites to use these new packages.
