+++ tests/cnf/ran/internal/version/version.go
@@ -1,4 +1,4 @@
-package ranhelper
+package version
 
 import (
 	"errors"
