+++ tests/cnf/ran/internal/version/version_test.go
@@ -1,4 +1,4 @@
-package ranhelper
+package version
 
 import (
 	"fmt"
